requirements.txt
# ============================================
# Drifting Oracle — Python Dependencies
# ============================================
# Install: pip install -r requirements.txt

# --- Core ---
pyspark==3.5.1
delta-spark==3.1.0
pandas>=2.0.0
numpy>=1.24.0

# --- ML & Experiment Tracking ---
scikit-learn>=1.3.0
mlflow>=2.12.0
xgboost>=2.0.0
shap>=0.43.0

# --- GenAI / RAG ---
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
openai>=1.14.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.1

# --- Visualization ---
matplotlib>=3.7.0
seaborn>=0.13.0

# --- Utilities ---
pyarrow>=14.0.0
